refactor(state): separate persisted runtime models
