+++ crates/horizon-core/src/runtime_state/tests/mod.rs
@@ -1,0 +1,8 @@
+mod panels;
+mod workspaces;
+
+use crate::board::WorkspaceLayout;
+use crate::config::{TerminalConfig, WorkspaceConfig};
+use crate::panel::{PanelOptions, PanelResume};
+
+use super::*;
