+++ docs/architecture/maintainability.md
@@ -62,9 +62,12 @@ back into large multi-purpose modules.
   `browser/manifest/request_queue.rs`, host-stamped workspace membership that
   scopes MCP discovery and control in `browser/manifest/workspace.rs`, and
   append-only audit storage in `browser/manifest/audit.rs`.
-- `runtime_state.rs` should stay focused on persisted board/window state; agent
-  binding orchestration, discovery, and external-store parsing belong in
-  `runtime_state/` helper modules. Binding validation and assignment live in
+- `runtime_state.rs` should stay focused on persisted board/window orchestration.
+  Persisted workspace, panel, template, and session-binding models live in
+  `runtime_state/models.rs`, with board/workspace and panel persistence tests in
+  `runtime_state/tests/`. Agent binding orchestration, discovery, and
+  external-store parsing belong in `runtime_state/` helper modules.
+  Binding validation and assignment live in
   `runtime_state/binding_bootstrap.rs`; provider-specific session-store parsing
   belongs in focused leaves such as `runtime_state/agent_sessions/codex.rs`.
 - `local_store.rs` centralizes agent-store environment paths and read-only
